requirements.txt
anthropic
streamlit
langgraph
langchain-anthropic
yfinance
python-dotenv